+++ CHANGES.rst
@@ -6,7 +6,8 @@ For changes before version 3.0, see ``HISTORY.rst``.
 5.4 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Prevent race condition in guarded_import
+  (see `#123 <https://github.com/zopefoundation/AccessControl/issues/123>`_)
 
 
 5.3 (2022-02-25)